+++ lib/presentation_layer/routes/home_page.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:camelus/presentation_layer/components/write_post.dart';
 import 'package:camelus/config/palette.dart';
-import 'package:camelus/presentation_layer/routes/nostr/nostr_drawer.dart';
+import 'package:camelus/presentation_layer/components/drawer/nostr_drawer.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:matomo_tracker/matomo_tracker.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
